fix: remove confusing timeouts that cancel early
Removes stream and websocket timeouts.
These were initially added with the idea of adding
a deadline for establishing a connection, however,
the implementation did not reflect this.
Contexts were being cancelled as soon as the connection
factory returns, meaning any calls made using the context
would fail, this was revealed in the implementation
of the celerity blueprint language server on json rpc
calls to the client when a change was made to stop
suppressing errors when those calls were made.